Fix small issues

+++ src/Randomizer.Core/DTOs/UpdateRoundResultDto.cs
@@ -1,9 +1,10 @@
 ﻿namespace Randomizer.Core.DTOs;
+
 public record UpdateRoundResultDto
 {
     public Guid Id { get; init; }
 
-    public int Score { get; init; }
+    public double Score { get; init; }
 
     public string? Comment { get; init; }
 }
